add fri state broadcaster

+++ kuka_sunrise_driver/include/kuka_sunrise/robot_control_client.hpp
@@ -85,7 +85,7 @@ private:
 
   bool torque_command_mode_ = false;
   double tracking_performance_;
-  int fri_state_;
+  double fri_state_;
 };
 
 }  // namespace kuka_sunrise
